Do plain implementation
